pyproject.toml
[project]
name = "league-ranker"
version = "0.0.1"
authors = [{ name = "Scott Houseman", email = "scott.houseman@gmail.com" }]
maintainers = [{ name = "Scott Houseman", email = "scott.houseman@gmail.com" }]
description = "A command-line application that will calculate the ranking table for a league"
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]

dependencies = ["click", "tabulate", "PyYAML"]

[project.urls]
Repository = "https://github.com/houseman/league-ranker"

[project.optional-dependencies]

dev = [
  # Developer tools for type-checking, formating, linting etc.
  "pre-commit",
  "black",
  "ruff",
  "mypy",

  # Mypy types
  "types-tabulate",
  "types-PyYAML",

  # Test dependencies
  "pytest",
  "pytest-cov[all]",
  "pytest-mock",
]

[project.scripts]
rank = "ranker.main:cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ranker*"]
namespaces = false

[tool.ruff]
src = ["src"]
select = ["E", "F", "B", "W", "I", "N", "D"]
ignore = [
  "D104", # Missing docstring in public package
  "D107", # Missing docstring in `__init__`
  "D203", # 1 blank line required before class docstring
  "D212", # Multi-line docstring summary should start at the first line
  "D401", # First line of docstring should be in imperative mood
]

# Same as Black.
line-length = 88

target-version = "py311"

[tool.ruff.per-file-ignores]
"**/tests/*" = [
  "D205", # 1 blank line required between summary line and description
  "D400", # First line should end with a period
  "D415", # First line should end with a period, question mark, or exclamation point
]

[tool.mypy]
mypy_path = ["src"]
python_version = 3.11
ignore_errors = false
disallow_untyped_calls = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_no_return = true

[[tool.mypy.overrides]]
module = ["tests.*"]
ignore_errors = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov --cov-report html --cov-report term-missing --cov-report xml --no-cov-on-fail --cov-fail-under 100"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 100
skip_empty = true
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if t.TYPE_CHECKING:",
  "pass",
  "raise NotImplementedError()",
]
